requirements.txt
fastapi
uvicorn
google-genai
google-adk
python-multipart
jinja2